refactor: rename observable.rs to from_fn.rs
